Remove postgres dependency
lava-server needs access to DB but not actual DB package installed
on the same machine. Removing dependency on postgres package to allow
for this.
Note that this dependency serves no purpose at present because
if postgres is not already installed *and* running by the time
lava-server gets installed, the installation fails. So pulling
postgres in dependencies  causes the install of lava-server to
fail